change logics of api'api/attractions' and logics of create DB program and adjust details present of front-end
