+++ .env
@@ -1,2 +1,6 @@
-config = '{"host" : "localhost","username" : "root","password" : "fuck_taipei" , "database" : "website_taipei"}'
-# config = '{"host" : "localhost","username" : "root","password" : "12345678" , "database" : "website_taipei"}'
+HOST = "localhost"
+# username 在 module "os" 好像是保留字
+DB_USERNAME = "root"
+DB_PASSWORD = "fuck_taipei"
+DATABASE = "website_taipei"
+DB_PASSWORD_BACKUP = "12345678"
