+++ app_folder/css/index.css
@@ -80,6 +80,11 @@ body{
     background: url("/static/images/Search.png") 50% 50% / contain no-repeat;
 }
 
+/* function */
+.pointer{
+    cursor: pointer;
+}
+
 /* Nav */
 .nevigation{
     display: flex;
@@ -198,7 +203,6 @@ body{
 .searchIcon{
     width: 30px;
     height: 30px;
-    cursor: pointer;
 }
 
 /* Main */
@@ -213,7 +217,7 @@ body{
     padding: 40px 0px;
     width: 100%;
     max-width: 1200px;
-    min-height: 800px;
+    min-height: calc(100vh - 478px - 80px);
     flex-direction: column;
     align-items: center;
     align-self: stretch;
@@ -265,16 +269,15 @@ body{
 .listItem{
     display: inline-block;
     padding: 5px 15px;
-    gap: 4px;
     color: #666;
     text-align: center;
     height: 14px;
-    cursor: pointer;
     
     white-space: nowrap;
 }
 .listItem:hover{
     color: black;
+    font-size: 20px;
 }
 .rightContainer{
     display: flex;
@@ -321,9 +324,11 @@ body{
     display: flex;
     flex-direction: column;
     justify-content: flex-start;
-    cursor: pointer;
 
     flex-basis: auto;
+}
+.attraction:hover{
+    opacity: 0.7;
 }
 @media(max-width: 1200px){
     .attraction{
