--- init_program/.env
@@ -1,1 +1,0 @@
-config = '{"host" : "localhost","username" : "root","password" : "fuck_taipei",}'
