+++ module/get_connection.py
@@ -1,3 +1,5 @@
+import os
+from dotenv import load_dotenv
 from flask import current_app
 import mysql.connector
 
@@ -9,7 +11,6 @@ def get_connection(func) -> bool:
             db_pool = current_app.config['connection_pool']
             db_connection = db_pool.get_connection()
             db_cursor = db_connection.cursor(dictionary=True)
-            print("OK")
             return func(db_connection, db_cursor, paras)
         except Exception as err:
             print(err)
@@ -21,6 +22,27 @@ def get_connection(func) -> bool:
     return wrapper
 
 
+# 建立 pool
 def create_pool(db_config):
     db_connection_pool = mysql.connector.pooling.MySQLConnectionPool(pool_name="pool_1", pool_size=5, **db_config)
     return db_connection_pool
+
+
+# 建立 與 DB 的連線(嘗試密碼)
+def access_db():
+    load_dotenv()
+    try:
+        db_config = {
+            "host": os.getenv("HOST"),
+            "username": os.getenv("DB_USERNAME"),
+            "password": os.getenv("DB_PASSWORD"),
+            "database": os.getenv("DATABASE"),
+        }
+        return create_pool(db_config)
+    except mysql.connector.errors.ProgrammingError as err:
+        if err.errno != 1045:
+            return
+        db_config["password"] = os.getenv("DB_PASSWORD_BACKUP")
+        return create_pool(db_config)
+    except Exception as err:
+        print(err)
